refactor: UserResponseDTO와 UserResponse 클래스명 변경
