+++ CatchMate/data/src/main/java/com/catchmate/data/datasource/remote/SignUpService.kt
@@ -2,7 +2,7 @@ package com.catchmate.data.datasource.remote
 
 import com.catchmate.data.dto.CheckNicknameResponseDTO
 import com.catchmate.data.dto.UserAdditionalInfoRequestDTO
-import com.catchmate.data.dto.UserResponseDTO
+import com.catchmate.data.dto.UserAdditionalInfoResponseDTO
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -18,5 +18,5 @@ interface SignUpService {
     @POST("user/additional-info")
     suspend fun postUserAdditionalInfo(
         @Body userAdditionalInfoRequestDTO: UserAdditionalInfoRequestDTO,
-    ): Response<UserResponseDTO?>
+    ): Response<UserAdditionalInfoResponseDTO?>
 }
