+++ CatchMate/domain/src/main/java/com/catchmate/domain/usecase/SignUpUseCase.kt
@@ -2,7 +2,7 @@ package com.catchmate.domain.usecase
 
 import com.catchmate.domain.model.CheckNicknameResponse
 import com.catchmate.domain.model.UserAdditionalInfoRequest
-import com.catchmate.domain.model.UserResponse
+import com.catchmate.domain.model.UserAdditionalInfoResponse
 import com.catchmate.domain.repository.SignUpRepository
 import javax.inject.Inject
 
@@ -13,6 +13,6 @@ class SignUpUseCase
     ) {
         suspend fun getNicknameAvailability(nickName: String): CheckNicknameResponse? = signUpRepository.getNicknameAvailability(nickName)
 
-        suspend fun postUserAdditionalInfo(userAdditionalInfoRequest: UserAdditionalInfoRequest): UserResponse? =
+        suspend fun postUserAdditionalInfo(userAdditionalInfoRequest: UserAdditionalInfoRequest): UserAdditionalInfoResponse? =
             signUpRepository.postUserAdditionalInfo(userAdditionalInfoRequest)
     }
